maihealth-api/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
django-shell-plus==1.1.7
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
